backend/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.12
exceptiongroup==1.2.2
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kafka-python==2.0.3
MarkupSafe==3.0.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
scikit-surprise==1.1.4
scipy==1.13.1
six==1.17.0
threadpoolctl==3.5.0
tomli==2.2.1
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
